// src/app/model/user.ts
export class User {
    // username: string;

    Password: string;
    
    // firstName: string;
    // lastName: string;

     PersonId: string;
    // BranchId: string;
    Email: string;
    // Code: string;
    // DocTipo: number;
    // DocNro: number;
    // Condicion: number;
    // TipoPersona: number;
    Denominacion: string;
    Cuit: string;
    // PtoVta: string;
    // IdProvincia: number;
    //Persona:
    // Direccion: string;
    // Localidad: string;
    // Cp: string;
    // Telefono1: string;
    // Telefono2: string;
    // EstablecimientoNro: string;
    // IngBrutos: string;
    // InicioActividad: string;
    // Estado: number;
    // PtoVtaEmitir: boolean;  

    AccessToken: string;
    Expiration: string; 

}